# Pertemuan 4/project_mysql/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TokoBersama API",
    "version": "1.0.0",
    "description": "REST API untuk platform marketplace TokoBersama. Mengelola kategori produk, seller, produk, dan order."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
...
  "components": {
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "message": { "type": "string" },
          "data": {}
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "message": { "type": "string" },
          "data": { "nullable": true, "example": null }
        }
      },
      "Kategori": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "nama_kategori": { "type": "string", "example": "Elektronik" },
          "deskripsi": {
            "type": "string",
            "example": "Produk elektronik seperti hp, laptop, dan TV"
          },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "KategoriInput": {
        "type": "object",
        "required": ["nama_kategori"],
        "properties": {
          "nama_kategori": { "type": "string", "example": "Elektronik" },
          "deskripsi": {
            "type": "string",
            "example": "Produk elektronik seperti hp, laptop, dan TV"
          }
        }
      },
      "Seller": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "nama_toko": { "type": "string", "example": "TechZone Official" },
          "email": { "type": "string", "example": "techzone@mail.com" },
          "no_hp": { "type": "string", "example": "081234567890" },
          "alamat": {
            "type": "string",
            "example": "Jl. Sudirman No. 10, Jakarta"
          },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "SellerInput": {
        "type": "object",
        "required": ["nama_toko", "email"],
        "properties": {
          "nama_toko": { "type": "string", "example": "TechZone Official" },
          "email": { "type": "string", "example": "techzone@mail.com" },
          "no_hp": { "type": "string", "example": "081234567890" },
          "alamat": {
            "type": "string",
            "example": "Jl. Sudirman No. 10, Jakarta"
          }
        }
      },
      "Produk": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "nama": { "type": "string", "example": "Samsung Galaxy A55" },
          "harga": { "type": "number", "example": 4999000 },
          "stok": { "type": "integer", "example": 50 },
          "id_kategori": { "type": "integer", "example": 1 },
          "id_seller": { "type": "integer", "example": 1 },
          "nama_kategori": { "type": "string", "example": "Elektronik" },
          "nama_toko": { "type": "string", "example": "TechZone Official" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "ProdukInput": {
        "type": "object",
        "required": ["nama", "harga", "id_kategori", "id_seller"],
        "properties": {
          "nama": { "type": "string", "example": "Samsung Galaxy A55" },
          "harga": { "type": "number", "example": 4999000 },
          "stok": { "type": "integer", "example": 50 },
          "id_kategori": { "type": "integer", "example": 1 },
          "id_seller": { "type": "integer", "example": 1 }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "id_produk": { "type": "integer", "example": 1 },
          "jumlah": { "type": "integer", "example": 2 },
          "total_harga": { "type": "number", "example": 9998000 },
          "status": {
            "type": "string",
            "enum": ["pending", "proses", "selesai", "batal"],
            "example": "pending"
          },
          "nama_produk": { "type": "string", "example": "Samsung Galaxy A55" },
          "harga_satuan": { "type": "number", "example": 4999000 },
          "nama_kategori": { "type": "string", "example": "Elektronik" },
          "nama_toko": { "type": "string", "example": "TechZone Official" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      }
    }
  },
  "paths": {
    "/kategori": {
      "get": {
        "tags": ["Kategori"],
        "summary": "Ambil semua kategori",
        "responses": {
          "200": {
            "description": "Daftar semua kategori",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/Kategori" }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Kategori"],
        "summary": "Tambah kategori baru",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/KategoriInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Kategori berhasil ditambahkan",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "properties": {
                        "data": { "$ref": "#/components/schemas/Kategori" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validasi gagal",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/kategori/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": { "type": "integer" },
          "example": 1
        }
      ],
      "get": {
        "tags": ["Kategori"],
        "summary": "Ambil kategori berdasarkan ID",
        "responses": {
          "200": {
            "description": "Data kategori",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "properties": {
                        "data": { "$ref": "#/components/schemas/Kategori" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Kategori tidak ditemukan",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Kategori"],
        "summary": "Update kategori",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/KategoriInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Kategori berhasil diupdate",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "properties": {
                        "data": { "$ref": "#/components/schemas/Kategori" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validasi gagal",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Kategori tidak ditemukan",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Kategori"],
        "summary": "Hapus kategori",
        "description": "Kategori tidak dapat dihapus jika masih memiliki produk terkait.",
        "responses": {
          "200": {
            "description": "Kategori berhasil dihapus",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          },
          "400": {
            "description": "Kategori masih memiliki produk",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Kategori tidak ditemukan",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/seller": {
      "get": {
        "tags": ["Seller"],
        "summary": "Ambil semua seller",
        "responses": {
          "200": {
            "description": "Daftar semua seller",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/Seller" }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Seller"],
        "summary": "Daftarkan seller baru",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SellerInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Seller berhasil didaftarkan",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "properties": {
                        "data": { "$ref": "#/components/schemas/Seller" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validasi gagal / email duplikat",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
    "/produk": {
      "get": {
        "tags": ["Produk"],
        "summary": "Ambil semua produk",
        "description": "Mendukung filter `?kategori=:id` dan pencarian `?search=:keyword`. Hasil sudah termasuk nama_kategori dan nama_toko (JOIN).",
        "parameters": [
          {
            "name": "kategori",
            "in": "query",
            "required": false,
            "schema": { "type": "integer" },
            "description": "Filter produk berdasarkan ID kategori",
            "example": 1
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": { "type": "string" },
            "description": "Cari produk berdasarkan nama (LIKE)",
            "example": "samsung"
          }
        ],
        "responses": {
          "200": {
            "description": "Daftar produk",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/Produk" }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Produk"],
        "summary": "Tambah produk baru",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProdukInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Produk berhasil ditambahkan",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "properties": {
                        "data": { "$ref": "#/components/schemas/Produk" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validasi gagal / FK tidak ditemukan",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
    "/order": {
      "get": {
        "tags": ["Order"],
        "summary": "Ambil semua order beserta detail produk",
        "responses": {
          "200": {
            "description": "Daftar semua order",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/Order" }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Order"],
        "summary": "Buat order baru",
        "description": "Menggunakan **MySQL Transaction**. Otomatis mengurangi stok produk dan menghitung `total_harga = harga × jumlah`. Menggunakan `SELECT ... FOR UPDATE` untuk mencegah race condition.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["id_produk", "jumlah"],
                "properties": {
                  "id_produk": { "type": "integer", "example": 1 },
                  "jumlah": { "type": "integer", "minimum": 1, "example": 2 }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order berhasil dibuat",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "properties": {
                        "data": { "$ref": "#/components/schemas/Order" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Stok tidak mencukupi / validasi gagal",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Produk tidak ditemukan",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  }
}